requirements.txt
Pillow>=9.0.0
ExifRead>=2.3.0